+++ src/main/java/com/zeylin/runoncoffee/controllers/dictionary/FoodItemController.java
@@ -1,13 +1,13 @@
 package com.zeylin.runoncoffee.controllers.dictionary;
 
-import com.zeylin.runoncoffee.models.dictionary.Dairy;
-import com.zeylin.runoncoffee.models.dictionary.Grains;
-import com.zeylin.runoncoffee.models.dictionary.Plant;
-import com.zeylin.runoncoffee.models.dictionary.Protein;
+import com.zeylin.runoncoffee.dto.dictionary.BaseFoodItemDto;
+import com.zeylin.runoncoffee.dto.dictionary.BaseFoodItemWithServingDto;
+import com.zeylin.runoncoffee.models.dictionary.base.BaseFoodItem;
 import com.zeylin.runoncoffee.services.dictionary.DairyService;
 import com.zeylin.runoncoffee.services.dictionary.GrainsService;
 import com.zeylin.runoncoffee.services.dictionary.PlantService;
 import com.zeylin.runoncoffee.services.dictionary.ProteinService;
+import org.modelmapper.ModelMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,7 +16,11 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
+/**
+ * Food items dictionaries.
+ */
 @RestController
 @RequestMapping("/dictionary/item")
 public class FoodItemController {
@@ -25,16 +29,19 @@ public class FoodItemController {
     private DairyService dairyService;
     private ProteinService proteinService;
     private PlantService plantService;
+    private ModelMapper modelMapper;
 
     public FoodItemController(GrainsService grainsService,
                               DairyService dairyService,
                               ProteinService proteinService,
-                              PlantService plantService
+                              PlantService plantService,
+                              ModelMapper modelMapper
     ) {
         this.grainsService = grainsService;
         this.dairyService = dairyService;
         this.proteinService = proteinService;
         this.plantService = plantService;
+        this.modelMapper = modelMapper;
     }
 
     /**
@@ -42,8 +49,11 @@ public class FoodItemController {
      */
     @GetMapping(value = "/grains", produces = "application/json")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<List<Grains>> getGrainItems() {
-        return ResponseEntity.ok(grainsService.getAll());
+    public ResponseEntity<List<BaseFoodItemWithServingDto>> getGrainItems() {
+        return ResponseEntity.ok(grainsService.getAll()
+                .stream()
+                .map(item -> convertToBaseFoodItemWithServingDto(item))
+                .collect(Collectors.toList()));
     }
 
     /**
@@ -51,8 +61,11 @@ public class FoodItemController {
      */
     @GetMapping(value = "/plants", produces = "application/json")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<List<Plant>> getPlantItems() {
-        return ResponseEntity.ok(plantService.getAll());
+    public ResponseEntity<List<BaseFoodItemDto>> getPlantItems() {
+        return ResponseEntity.ok(plantService.getAll()
+                .stream()
+                .map(item -> convertToBaseFoodItemDto(item))
+                .collect(Collectors.toList()));
     }
 
     /**
@@ -60,8 +73,11 @@ public class FoodItemController {
      */
     @GetMapping(value = "/dairy", produces = "application/json")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<List<Dairy>> getDairyItems() {
-        return ResponseEntity.ok(dairyService.getAll());
+    public ResponseEntity<List<BaseFoodItemWithServingDto>> getDairyItems() {
+        return ResponseEntity.ok(dairyService.getAll()
+                .stream()
+                .map(item -> convertToBaseFoodItemWithServingDto(item))
+                .collect(Collectors.toList()));
     }
 
     /**
@@ -69,8 +85,19 @@ public class FoodItemController {
      */
     @GetMapping(value = "/protein", produces = "application/json")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<List<Protein>> getProteinItems() {
-        return ResponseEntity.ok(proteinService.getAll());
+    public ResponseEntity<List<BaseFoodItemWithServingDto>> getProteinItems() {
+        return ResponseEntity.ok(proteinService.getAll()
+                .stream()
+                .map(item -> convertToBaseFoodItemWithServingDto(item))
+                .collect(Collectors.toList()));
+    }
+
+    private BaseFoodItemDto convertToBaseFoodItemDto(BaseFoodItem item) {
+        return modelMapper.map(item, BaseFoodItemDto.class);
+    }
+
+    private BaseFoodItemWithServingDto convertToBaseFoodItemWithServingDto(BaseFoodItem item) {
+        return modelMapper.map(item, BaseFoodItemWithServingDto.class);
     }
 
 }