+++ src/main/java/com/zeylin/runoncoffee/dto/dailyrecord/DailyRecordSaveDto.java
@@ -1,19 +1,16 @@
-package com.zeylin.runoncoffee.dto;
+package com.zeylin.runoncoffee.dto.dailyrecord;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
-import java.time.LocalDate;
-
 @Getter
 @Setter
 @NoArgsConstructor
 @ToString
-public class DailyRecordDisplayDto {
+public class DailyRecordSaveDto {
 
-    private LocalDate day;
     private Integer grains;
     private Integer veggie;
     private Integer dairy;