+++ src/main/java/com/zeylin/runoncoffee/dto/dailyrecord/DailyRecordUpdateDto.java
@@ -1,4 +1,4 @@
-package com.zeylin.runoncoffee.dto;
+package com.zeylin.runoncoffee.dto.dailyrecord;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;