+++ src/main/java/com/zeylin/runoncoffee/dto/dictionary/BaseFoodItemDto.java
@@ -1,0 +1,13 @@
+package com.zeylin.runoncoffee.dto.dictionary;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class BaseFoodItemDto {
+    private Long id;
+    private String name;
+}