+++ src/main/java/com/zeylin/runoncoffee/dto/dictionary/BaseFoodItemWithServingDto.java
@@ -1,0 +1,12 @@
+package com.zeylin.runoncoffee.dto.dictionary;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class BaseFoodItemWithServingDto extends BaseFoodItemDto {
+    private String servingName;
+}