+++ src/test/java/com/zeylin/runoncoffee/services/DailyRecordServiceTest.java
@@ -1,9 +1,9 @@
 package com.zeylin.runoncoffee.services;
 
-import com.zeylin.runoncoffee.dto.DailyRecordAveragesDto;
-import com.zeylin.runoncoffee.dto.DailyRecordDisplayDto;
-import com.zeylin.runoncoffee.dto.DailyRecordSaveDto;
-import com.zeylin.runoncoffee.dto.DailyRecordStatsDto;
+import com.zeylin.runoncoffee.dto.dailyrecord.DailyRecordAveragesDto;
+import com.zeylin.runoncoffee.dto.dailyrecord.DailyRecordDisplayDto;
+import com.zeylin.runoncoffee.dto.dailyrecord.DailyRecordSaveDto;
+import com.zeylin.runoncoffee.dto.dailyrecord.DailyRecordStatsDto;
 import com.zeylin.runoncoffee.exceptions.NotFoundException;
 import com.zeylin.runoncoffee.models.DailyRecord;
 import com.zeylin.runoncoffee.repositories.DailyRecordRepository;