+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.4.0-SNAPSHOT] - 2019-09-10
+### Added
+- Return DTOs for food item dictionaries
+
+### Changed
+- Moved DailyRecord DTOs to a separate package
+
 ## [0.3.0-SNAPSHOT] - 2019-09-06
 ### Added
 - Dictionaries for food items by group